refactor/ using BaseRequest for same rule application
