+++ app/Http/Requests/User/Order/OrderShippingListingRequest.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Requests\User\Order;
 
-use Illuminate\Foundation\Http\FormRequest;
+use App\Http\Requests\BaseRequest;
 
-class OrderShippingListingRequest extends FormRequest
+class OrderShippingListingRequest extends BaseRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -21,17 +21,13 @@ class OrderShippingListingRequest extends FormRequest
      */
     public function rules(): array
     {
-        return [
-            'page' => 'nullable|integer',
-            'limit' => 'nullable|integer',
-            'sortBy' => 'nullable|string',
-            'desc' => 'nullable|boolean',
+        return array_merge(parent::rules(), [
             'dateRange' => 'nullable|array',
             'dateRange.from' => 'nullable|date|before_or_equal:dateRange.to',
             'dateRange.to' => 'nullable|date|after_or_equal:dateRange.from',
             'fixRange' => 'nullable|in:today,monthly,yearly',
             'orderUuid' => 'nullable|string|exists:orders,uuid',
             'customerUuid' => 'nullable|string|exists:users,uuid',
-        ];
+        ]);
     }
 }
