+++ app/Models/Post.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Builders\Post\PostBuilder;
+use App\Traits\HasImage;
 use App\Traits\Sluggable;
 use App\Traits\Uuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -10,7 +11,7 @@ use Illuminate\Database\Eloquent\Model;
 
 class Post extends Model
 {
-    use HasFactory, Uuids, Sluggable;
+    use HasFactory, Uuids, Sluggable, HasImage;
 
     /**
      * The attributes that are mass assignable.
@@ -33,25 +34,6 @@ class Post extends Model
     protected $casts = [
         'metadata' => 'array',
     ];
-
-    /**
-     * Get the image uuid of the metadata attribute.
-     * @return string
-     */
-    public function getImageUuidAttribute(): string
-    {
-        return $this->metadata['image'];
-    }
-
-    /**
-     * Get the related image.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
-     */
-    public function image()
-    {
-        return $this->belongsTo(File::class, 'image_uuid', 'uuid');
-    }
 
     /**
      * Instantiate a new QueryBuilder instance.
