+++ app/Models/Promotion.php
@@ -3,13 +3,14 @@
 namespace App\Models;
 
 use App\Builders\Promotion\PromotionBuilder;
+use App\Traits\HasImage;
 use App\Traits\Uuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Promotion extends Model
 {
-    use HasFactory, Uuids;
+    use HasFactory, Uuids, HasImage;
 
     /**
      * The attributes that are mass assignable.
@@ -31,25 +32,6 @@ class Promotion extends Model
     protected $casts = [
         'metadata' => 'array',
     ];
-
-    /**
-     * Get the image uuid of the metadata attribute.
-     * @return string
-     */
-    public function getImageUuidAttribute(): string
-    {
-        return $this->metadata['image'];
-    }
-
-    /**
-     * Get the related image.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
-     */
-    public function image()
-    {
-        return $this->belongsTo(File::class, 'image_uuid', 'uuid');
-    }
 
     /**
      * Instantiate a new QueryBuilder instance.
